requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.8
django-allauth==0.45.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-dotenv==1.4.2
django-rest-auth==0.9.5
django-timezone-field==4.2.1
djangorestframework==3.12.4
idna==3.2
kombu==5.1.0
oauthlib==3.1.1
prompt-toolkit==3.0.20
pycparser==2.20
PyJWT==2.2.0
python-crontab==2.5.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
redis==3.5.3
redis-server==6.0.9
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
